moved genInitReductionTargets from BackendSIMT to BackendBase
