+++ include/genn/genn/code_generator/backendSIMT.h
@@ -220,23 +220,6 @@ protected:
 
 private:
     //--------------------------------------------------------------------------
-    // ReductionTarget
-    //--------------------------------------------------------------------------
-    //! Simple struct to hold reduction targets
-    struct ReductionTarget
-    {
-        ReductionTarget(const std::string &n, const std::string &t, VarAccessMode a)
-        : name(n), type(t), access(a)
-        {
-        }
-
-        const std::string name;
-        const std::string type;
-        const VarAccessMode access;
-    };
-
-
-    //--------------------------------------------------------------------------
     // Type definitions
     //--------------------------------------------------------------------------
     template<typename T>
@@ -330,31 +313,6 @@ private:
     {
         genParallelGroup(os, kernelSubs, groups, idStart, getPaddedSizeFunc,
                          [](const T &) { return true; }, handler);
-    }
-    
-    template<typename G>
-    std::vector<ReductionTarget> genInitReductionTargets(CodeStream &os, const G &cg) const
-    {
-        // Loop through variables
-        std::vector<ReductionTarget> reductionTargets;
-        const auto *cm = cg.getArchetype().getCustomUpdateModel();
-        for(const auto &v : cm->getVars()) {
-            // If variable is a reduction target, define variable initialised to correct initial value for reduction
-            if(v.access & VarAccessModeAttribute::REDUCE) {
-                os << v.type << " lr" << v.name << " = " << getReductionInitialValue(*this, getVarAccessMode(v.access), v.type) << ";" << std::endl;
-                reductionTargets.emplace_back(v.name, v.type, getVarAccessMode(v.access));
-            }
-        }
-
-        // Loop through variable references
-        for(const auto &v : cm->getVarRefs()) {
-            // If variable reference is a reduction target, define variable initialised to correct initial value for reduction
-            if(v.access & VarAccessModeAttribute::REDUCE) {
-                os << v.type << " lr" << v.name << " = " << getReductionInitialValue(*this, v.access, v.type) << ";" << std::endl;
-                reductionTargets.emplace_back(v.name, v.type, v.access);
-            }
-        }
-        return reductionTargets;
     }
     
     // Helper function to generate kernel code to initialise variables associated with synapse group or custom WU update with dense/kernel connectivity
